Readme update, adjustment to monitor_service_blueiris.cpp to fix compile issues on arm

+++ lib/monitor_service_blueiris.cpp
@@ -192,7 +192,7 @@ MonitorServiceBlueIris::imageDimensions(const char *image, unsigned int length) 
   // We'll read those 12 bytes at buf+2...buf+14, i.e. overwriting the existing buf.
   if (buf[0]==0xFF && buf[1]==0xD8 && buf[2]==0xFF && buf[3]==0xE0 && 
 		buf[6]=='J' && buf[7]=='F' && buf[8]=='I' && buf[9]=='F') { 
-		long pos=2;
+		unsigned int pos=2;
     while (buf[2]==0xFF) { 
 			if (buf[3]==0xC0 || buf[3]==0xC1 || buf[3]==0xC2 || buf[3]==0xC3 || 
 				buf[3]==0xC9 || buf[3]==0xCA || buf[3]==0xCB)
